+++ web_service/src/errors.rs
@@ -1,12 +1,17 @@
 use axum::response::{IntoResponse, Response};
 
+pub type Result<T> = std::result::Result<T, Error>;
+
 #[derive(thiserror::Error, Debug)]
-pub enum WebError {
+pub enum Error {
     #[error(transparent)]
-    GrpcError(tonic::Status),
+    GrpcStubError(#[from] tonic::Status),
+
+    #[error(transparent)]
+    GrpcConnectionError(#[from] tonic::transport::Error)
 }
 
-impl IntoResponse for WebError {
+impl IntoResponse for Error {
     fn into_response(self) -> Response {
         todo!()
     }
