+++ web_service/src/routes/mod.rs
@@ -1,7 +1,6 @@
 use axum::{Router, routing::get};
 use models::list_models;
-
-use crate::AppState;
+use crate::server::AppState;
 
 mod chat;
 mod embed;
@@ -11,7 +10,7 @@ async fn hello() -> &'static str {
     "hello, world"
 }
 
-pub fn app_routes() -> Router<AppState> {
+pub(crate) fn app_routes() -> Router<AppState> {
     let model_routes = Router::new()
         .route("/list", get(list_models));
         // .route("/{id}/chat", todo!()) // not married to this routing ...
